bubble game
The bubble game. It's a scoring game. Like in a game, you have a time limit. to find a number and click on a number. Which number you find, and they will increase your time limit. and if you do not find out the number, they will be put on hold. and you restart the game. This game. Make use of basic concepts like HTML, CSS, and JavaScript.